Simplifying SqlStatementClient by introducing "mother-parser"
