+++ source/dotnet/wholesale-api/WebApi/Configuration/Modules/CalculationResultsRegistration.cs
@@ -36,6 +36,7 @@ public static class CalculationResultsRegistration
         serviceCollection.AddScoped<ISqlStatementClient, SqlStatementClient>();
         serviceCollection.AddScoped<ISettlementReportResultsCsvWriter, SettlementReportResultsCsvWriter>();
         serviceCollection.AddScoped<IDatabricksSqlResponseParser, DatabricksSqlResponseParser>();
+        serviceCollection.AddScoped<IDatabricksSqlStatusResponseParser, DatabricksSqlStatusResponseParser>();
         serviceCollection.AddScoped<IDatabricksSqlChunkResponseParser, DatabricksSqlChunkResponseParser>();
         serviceCollection.AddScoped<IDatabricksSqlChunkDataResponseParser, DatabricksSqlChunkDataResponseParser>();
         serviceCollection.AddScoped<IDataLakeClient, DataLakeClient>();
